Ajuste no id do arquivo

+++ src/files/files.service.ts
@@ -11,14 +11,13 @@ import {
 import { Company } from './entities/company.entity';
 import { Program } from './entities/program.entity';
 import { UploadFileDto } from './dto/upload-file.dto';
-// import { getSignedUrl } from '@aws-sdk/s3-request-presigner';
 
-// import { Company } from './entities/company.entity';
 export interface FolderNode {
   name: string;
   subfolders: FolderNode[];
   files: any;
 }
+
 @Injectable()
 export class FilesService {
   private s3: AWS.S3;
@@ -39,7 +38,7 @@ export class FilesService {
     });
 
     this.s3Client = new S3Client({
-      region: process.env.AWS_REGION, // Exemplo: 'us-east-1'
+      region: process.env.AWS_REGION,
       credentials: {
         accessKeyId: process.env.AWS_ACCESS_KEY_ID,
         secretAccessKey: process.env.AWS_SECRET_ACCESS_KEY,
@@ -47,18 +46,17 @@ export class FilesService {
     });
   }
 
+  // Upload de arquivos para o S3
   public async uploadFile(fileName: string, file: Buffer, uploadFileDto: UploadFileDto) {
     const program = await this.programRepository.findOne({
-      where: {
-        id: uploadFileDto.programId,
-      },
+      where: { id: uploadFileDto.programId },
     });
+    if (!program) throw new Error('Programa não encontrado');
 
     const company = await this.companyRepository.findOne({
-      where: {
-        id: program.companyId,
-      },
+      where: { id: program.companyId },
     });
+    if (!company) throw new Error('Empresa não encontrada');
 
     const cnpj = company.cnpj.replace(/[.\-\/]/g, '');
     const bucketName = process.env.AWS_S3_BUCKET_NAME;
@@ -80,85 +78,49 @@ export class FilesService {
       file_logo_id: 1,
       url: `${key}`,
     });
-    return program;
+
+    return { message: 'Arquivo enviado com sucesso', program };
   }
+
+  // Criação de pastas específicas com subpastas
   async createFolder(ldb: string, folderName: string = null) {
-    if (folderName === 'CONTÁBIL') {
-      const subFolders = [
-        'Controles internos',
-        'Notas Fiscais',
-        'RH',
-        'Timesheet',
-        'Valoração',
-      ];
+    const folderMap = {
+      'CONTÁBIL': ['Controles internos', 'Notas Fiscais', 'RH', 'Timesheet', 'Valoração'],
+      'TÉCNICO': ['Anexo FORMP&D', 'Arquivado', 'Documentação Técnica', 'Imagens'],
+      'ENTREGÁVEL': ['Dossiê', 'FORMP&D', 'Parecer MCTI', 'Relatório Gerencial'],
+    };
 
-      for (const subFolder of subFolders) {
-        const params = {
-          Bucket: process.env.AWS_S3_BUCKET_NAME,
-          Key: `${ldb}/CONTÁBIL/${subFolder}/`, // Note the trailing slash to create a folder
-          Body: '',
-        };
-        await this.s3.putObject(params).promise();
-      }
-    } else if (folderName === 'TÉCNICO') {
-      const subFolders = [
-        'Anexo FORMP&D',
-        'Arquivado',
-        'Documentação Técnica',
-        'Imagens',
-      ];
+    const subFolders = folderMap[folderName] || [];
 
-      for (const subFolder of subFolders) {
-        const params = {
-          Bucket: process.env.AWS_S3_BUCKET_NAME,
-          Key: `${ldb}/TÉCNICO/${subFolder}/`, // Note the trailing slash to create a folder
-          Body: '',
-        };
-        await this.s3.putObject(params).promise();
-      }
-    } else {
-      const subFolders = [
-        'Dossiê',
-        'FORMP&D',
-        'Parecer MCTI',
-        'Relatório Gerencial',
-      ];
-
-      for (const subFolder of subFolders) {
-        const params = {
-          Bucket: process.env.AWS_S3_BUCKET_NAME,
-          Key: `${ldb}/ENTREGÁVEL/${subFolder}/`, // Note the trailing slash to create a folder
-          Body: '',
-        };
-        await this.s3.putObject(params).promise();
-      }
+    for (const subFolder of subFolders) {
+      const params = {
+        Bucket: process.env.AWS_S3_BUCKET_NAME,
+        Key: `${ldb}/${folderName}/${subFolder}/`, // Trailing slash para criar a pasta
+        Body: '',
+      };
+      await this.s3.putObject(params).promise();
     }
   }
+
+  // Criação de múltiplas pastas com subpastas para cada programa
   async createFolders(body: any) {
-    
-    const company = await this.companyRepository.findOneBy({
-      id: body.companyId,
-    });
-    const programs = await this.programRepository.find({
-      where: {
-        companyId: company.id,
-      },
-    });
-    
+    const company = await this.companyRepository.findOneBy({ id: body.companyId });
+    const programs = await this.programRepository.find({ where: { companyId: company.id } });
 
     for (const program of programs) {
       const cnpj = company.cnpj.replace(/[.\-\/]/g, '');
       const ldb = `${cnpj}/${program.name}`;
       const folders = ['CONTÁBIL', 'ENTREGÁVEL', 'TÉCNICO'];
+
       for (const folder of folders) {
         await this.createFolder(ldb, folder);
       }
     }
 
     return programs;
-    
   }
 
+  // Listar todas as pastas e subpastas com seus arquivos
   async listAllFolders(programId: number) {
     const program = await this.programRepository.findOneBy({
       id: programId,
@@ -171,11 +133,11 @@ export class FilesService {
     const cnpj = company.cnpj.replace(/[.\-\/]/g, '');
 
     const folderPath = `${cnpj}/${program.name}`;
-    // const folderPath = 'ACESSO DIGITAL TECNOLOGIA DA INFORMACAO S.A.';
     const folders = await this.getAllSubfolders(folderPath);
     return folders.subfolders;
   }
 
+  // Obter todas as subpastas e arquivos de um diretório no S3
   async getAllSubfolders(folderName: string): Promise<FolderNode> {
     const bucketName = process.env.AWS_S3_BUCKET_NAME!;
 
@@ -190,83 +152,79 @@ export class FilesService {
       const subfolders =
         response.CommonPrefixes?.map((prefix) => prefix.Prefix) ?? [];
       const subfolderNodes = await Promise.all(subfolders.map(fetchSubfolders));
+
       const regex = /\/([^\/]+)\/?$/;
       const match = prefix.match(regex);
-      const files = await this.listFiles(prefix);
-      
+      const files = await this.listFiles(prefix); // Lista os arquivos com ID e URL
+
       return {
-        name: match[1],
+        name: match ? match[1] : 'root', // Nome da pasta
         subfolders: subfolderNodes,
-        files: files,
+        files: files, // Lista os arquivos com ID e URL
       };
     };
 
+    // Garante que sempre seja retornado um folderNode, mesmo se a pasta não tiver subpastas
     return await fetchSubfolders(
       folderName.endsWith('/') ? folderName : `${folderName}/`,
     );
   }
 
+  // Listar arquivos de uma subpasta
   async listFiles(folder: string) {
     const params = {
       Bucket: process.env.AWS_S3_BUCKET_NAME,
       Prefix: folder,
       Delimiter: '/',
     };
-  
+
     try {
-      const data = await this.s3.listObjectsV2(params).promise();
-      
+      const data = await this.s3Client.send(new ListObjectsV2Command(params));
+
+      if (!data || !data.Contents) {
+        // Retorna um array vazio se não houver conteúdo (nenhum arquivo)
+        return [];
+      }
+
       // Filtrar somente os arquivos que estão na raiz da pasta, sem subpastas
-      const files = data.Contents.map((item) => item.Key)
-        .filter((key) => !key.endsWith('/')) // Filtra apenas os arquivos
-        .map((key) => {
-          const match = key.match(/\/([^\/]+)\/?$/);
-          return match ? match[1] : null;
-        })
-        .filter(Boolean); // Remove valores nulos
-  
-      return files;
+      const files = data.Contents.filter((item) => !item.Key.endsWith('/')).map(async (item) => {
+        const key = item.Key;
+        const fileEntity = await this.filesRepository.findOne({ where: { url: key } });
+
+        return {
+          id: fileEntity ? fileEntity.id : null,
+          name: key ? key.split('/').pop() : null,
+          url: `https://${params.Bucket}.s3.amazonaws.com/${key}`, // URL completa do arquivo no S3
+        };
+      });
+
+      // Resolve as promessas e retorna os arquivos
+      return await Promise.all(files);
     } catch (err) {
       console.error('Erro ao listar os arquivos:', err);
-      throw err;
+      // Retorna um array vazio em caso de erro
+      return [];
     }
   }
-  
 
-  // async checkIfObjectExists(bucket, key) {
-  //   try {
-  //     await this.s3.headObject({ Bucket: bucket, Key: key }).promise();
-  //     console.log('Object exists');
-  //     return true;
-  //   } catch (error) {
-  //     if (error.code === 'NotFound') {
-  //       return false;
-  //     }
-  //     // Handle other errors, e.g., permission issues
-  //     throw error;
-  //   }
-  // }
+  // Deletar arquivos de uma pasta no S3 e banco de dados
   async delete(id: number) {
-    const file = await this.filesRepository.findOne({
-      where: {
-        id: id,
-      },
-    });
-    const bucketName = process.env.AWS_S3_BUCKET_NAME;
-    // const regex = /^https:\/\//i;
+    const file = await this.filesRepository.findOne({ where: { id } });
+    if (!file) throw new Error('Arquivo não encontrado');
 
     const params = {
-      Bucket: bucketName,
+      Bucket: process.env.AWS_S3_BUCKET_NAME,
       Key: file.url,
     };
 
     try {
       await this.s3.deleteObject(params).promise();
       await this.filesRepository.delete(file.id);
-      console.log(`File deleted successfully from `);
+      console.log(`Arquivo deletado com sucesso: ${file.url}`);
       return file;
     } catch (error) {
-      throw new Error(`Failed to delete file from S3: ${error.message}`);
+      console.error('Erro ao deletar arquivo do S3:', error);
+      throw new Error(`Erro ao deletar arquivo do S3: ${error.message}`);
     }
   }
 }
